Added gitignore file Fixed variable histogram to static uint16_t --- Problems find -> variable static uint16_t histogram have overflow with 255 values for unknown reason

+++ Lab2/imageFunctions.c
@@ -11,10 +11,14 @@ Desenvolvido para a placa EK-TM4C1294XL utilizando o SDK TivaWare no KEIL
 #include <stdint.h>
 #include "imageFunctions.h"
 
-uint16_t EightBitHistogram_C(uint16_t width, uint16_t height, const uint8_t *p_image, uint8_t *p_histogram)
+// Gera histograma para imagens até 64K. 
+// Caso a imagem seja maior que 64K retorna 0, caso contrario retorna o tamanho da imagem.
+// Usa o ponteiro p_histogram como retorno do histograma
+uint16_t EightBitHistogram_C(uint16_t width, uint16_t height, const uint8_t *p_image, uint16_t *p_histogram)
 {
+	// Variavel com o tamanho da imagem
 	uint32_t image_size = width*height;
-	
+	// Verifica se a imagem é maior que 64K
 	if(image_size < image_size_limit){
 			//clear histogram
 			for(uint16_t i=0;i<256;i++){
